Move `Version.createFromIngestion` logic.
