+++ models/version.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const cloneDeep = require('lodash/cloneDeep');
 const isPlainObject = require('lodash/isPlainObject');
 const path = require('path');
 const semver = require('semver');
@@ -8,6 +7,7 @@ const {removeStopwords} = require('stopword');
 const { v4: uuid, v5: uuidv5 } = require('uuid');
 const union = require('lodash/union');
 const propertyFilter = require('../lib/model-property-filter');
+const createVersionFromIngestion = require('../lib/create-version-from-ingestion');
 
 const origamiSupportEmail = 'origami.support@ft.com';
 
@@ -584,164 +584,7 @@ function initModel(app) {
 
 		// Create a version based on an Ingestion
 		async createFromIngestion(ingestion) {
-			try {
-				const url = ingestion.get('url');
-				const tag = ingestion.get('tag');
-				const encodedTag = encodeURIComponent(tag);
-
-				// Expect a valid GitHub URL
-				if (!app.github.isValidUrl(url)) {
-					throw app.github.error('Ingestion URL is not a GitHub repository', false);
-				}
-				const {owner, repo} = app.github.extractRepoFromUrl(url);
-
-				// Check that the repo/tag exist
-				const repoTagExists = await app.github.isValidRepoAndTag({
-					owner,
-					repo,
-					tag: encodedTag
-				});
-				if (!repoTagExists) {
-					throw app.github.error('Repo or tag does not exist', false);
-				}
-
-				// Load the Origami manifest first
-				const origamiManifest = await app.github.loadJsonFile({
-					path: 'origami.json',
-					ref: tag,
-					owner,
-					repo
-				});
-				if (!origamiManifest) {
-					throw app.github.error('Repo does not contain an Origami manifest', false);
-				}
-
-				// Extract the information we need from the Origami manifest
-				const origamiManifestNormalised = Version.normaliseOrigamiManifest(origamiManifest);
-
-				// Load the other manifests
-				const [aboutManifest, bowerManifest, imageSetManifest, packageManifest] = await Promise.all([
-					app.github.loadJsonFile({
-						path: 'about.json',
-						ref: tag,
-						owner,
-						repo
-					}),
-					app.github.loadJsonFile({
-						path: 'bower.json',
-						ref: tag,
-						owner,
-						repo
-					}),
-					app.github.loadJsonFile({
-						path: 'imageset.json',
-						ref: tag,
-						owner,
-						repo
-					}),
-					app.github.loadJsonFile({
-						path: 'package.json',
-						ref: tag,
-						owner,
-						repo
-					})
-				]);
-
-				// Load repo markdown
-				const readme = await app.github.loadReadme({
-					ref: tag,
-					owner,
-					repo
-				});
-				const designguidelines = await app.github.loadFile({
-					path: 'designguidelines.md',
-					ref: tag,
-					owner,
-					repo
-				});
-				const migration = await app.github.loadFile({
-					path: 'migration.md',
-					ref: tag,
-					owner,
-					repo
-				});
-
-				// Create the new version
-				const version = new Version({
-					name: repo,
-					type: origamiManifestNormalised.origamiType,
-					url,
-					tag,
-					support_status: origamiManifestNormalised.supportStatus,
-					support_email: origamiManifestNormalised.supportContact.email,
-					support_channel: origamiManifestNormalised.supportContact.slack,
-					manifests: {
-						about: aboutManifest,
-						bower: bowerManifest,
-						imageSet: imageSetManifest,
-						origami: origamiManifest,
-						package: packageManifest
-					},
-					markdown: {
-						readme,
-						designguidelines,
-						migration
-					}
-				});
-				await version.save();
-
-				// Return the new version
-				return version;
-
-			} catch (error) {
-				// Assume errors are recoverable by default
-				if (error.isRecoverable !== false) {
-					error.isRecoverable = true;
-				}
-				throw error;
-			}
-
-		},
-
-		// Normalise an Origami manifest file. Because we use the manifest
-		// to create fields in the database, we need to ensure that it
-		// mostly conforms to the Origami spec.
-		normaliseOrigamiManifest(manifest) {
-			const normalisedManifest = cloneDeep(manifest);
-
-			// Ensure that origamiType is a string or null
-			if (typeof normalisedManifest.origamiType !== 'string') {
-				normalisedManifest.origamiType = null;
-			}
-
-			// Convert the "component" origamiType to "module" in the database
-			if (normalisedManifest.origamiType === 'component') {
-				normalisedManifest.origamiType = 'module';
-			}
-
-			// Ensure that support status is a string or null
-			if (typeof normalisedManifest.support !== 'string') {
-				normalisedManifest.support = null;
-			}
-
-			// Ensure that we have all the support information that we need
-			if (typeof normalisedManifest.supportContact !== 'object' || normalisedManifest.supportContact === null) {
-				normalisedManifest.supportContact = {
-					email: null,
-					slack: null
-				};
-			}
-			if (!normalisedManifest.supportContact.email && typeof normalisedManifest.support === 'string' && normalisedManifest.support.includes('@')) {
-				normalisedManifest.supportContact.email = normalisedManifest.support;
-			}
-			if (!normalisedManifest.supportContact.email) {
-				normalisedManifest.supportContact.email = origamiSupportEmail;
-			}
-			if (!normalisedManifest.supportContact.slack && normalisedManifest.supportContact.email === origamiSupportEmail) {
-				normalisedManifest.supportContact.slack = 'financialtimes/origami-support';
-			}
-
-			return normalisedManifest;
+			await createVersionFromIngestion(ingestion, app.model.Version, app.github);
 		},
 
 		// Normalise an Origami manifest demo
